# PERF: accessing sliced indexes with populated indexing engines