--- a/pandas/_libs/index.pyx
+++ b/pandas/_libs/index.pyx
@@ -230,17 +230,15 @@ cdef class IndexEngine:
     def __sizeof__(self) -> int:
         return self.sizeof()
 
-    cpdef _update_from_other(self, IndexEngine other, reverse: bool):
+    cpdef _update_from_sliced(self, IndexEngine other, reverse: bool):
         self.unique = other.unique
         self.need_unique_check = other.need_unique_check
-        self.need_monotonic_check = other.need_monotonic_check
-        if reverse and not self.need_monotonic_check:
-            # the slice reverses the index
-            self.monotonic_inc = other.monotonic_dec
-            self.monotonic_dec = other.monotonic_inc
-        else:
-            self.monotonic_inc = other.monotonic_inc
-            self.monotonic_dec = other.monotonic_dec
+        if not other.need_monotonic_check and (
+                other.is_monotonic_increasing or other.is_monotonic_decreasing):
+            self.need_monotonic_check = other.need_monotonic_check
+            # reverse=True means the index has been reversed
+            self.monotonic_inc = other.monotonic_dec if reverse else other.monotonic_inc
+            self.monotonic_dec = other.monotonic_inc if reverse else other.monotonic_dec
 
     @property
     def is_unique(self) -> bool: