--- a/pandas/_libs/index.pyx
+++ b/pandas/_libs/index.pyx
@@ -902,6 +902,16 @@ cdef class SharedEngine:
         # for compat with IndexEngine
         pass
 
+    cpdef _update_from_sliced(self, ExtensionEngine other, reverse: bool):
+        self.unique = other.unique
+        self.need_unique_check = other.need_unique_check
+        if not other.need_monotonic_check and (
+                other.is_monotonic_increasing or other.is_monotonic_decreasing):
+            self.need_monotonic_check = other.need_monotonic_check
+            # reverse=True means the index has been reversed
+            self.monotonic_inc = other.monotonic_dec if reverse else other.monotonic_inc
+            self.monotonic_dec = other.monotonic_inc if reverse else other.monotonic_dec
+
     @property
     def is_unique(self) -> bool:
         if self.need_unique_check: