--- a/pandas/core/indexes/base.py
+++ b/pandas/core/indexes/base.py
@@ -5180,7 +5180,7 @@ class Index(IndexOpsMixin, PandasObject):
         """
         res = self._data[slobj]
         result = type(self)._simple_new(res, name=self._name)
-        if isinstance(result._engine, libindex.IndexEngine):  # may also be IntervalTree
+        if "_engine" in self._cache:
             reverse = slobj.step is not None and slobj.step < 0
             result._engine._update_from_sliced(self._engine, reverse=reverse)