@jbrockmendel 2023-03-02
nice! this has been on my todo list for ages but was always intimidating